Do not check for deep storage w/o LWM Deep Storage
Closes #3 hopefully!

There may be a better way to solve this, possibly including the
IHoldMultipleThings.dll along with this mod? Also some of what I thought
was "Deep Storage" vs "Vanilla" hauling behavior may be chalked up to
JobDefOf.HaulToCell vs JobDefOf.HaulToContainer jobs. Could possibly
find another function to patch to improve refilling of a partial stacks
in containers if that is the case.