bananasss00 Fix constructed storage defaults
Previously, if you had set-up and copied kanban stockpile settings to
clipboard, a newly constructed storage building would auto paste those
settings which was a bug.

All credits to bananasss00, but I was unable to merge commit as-is, as
it commits binary dll code which I choose not to store in git.

https://github.com/bananasss00/RimWorld-KanbanStockpile/commit/34ee4a48202e21858d43424b8e17d9bf5c9c741b